+++ R/checkString.r
@@ -8,8 +8,7 @@
 #' @template na-handling
 #' @param na.ok [\code{logical(1)}]\cr
 #'  Are missing values allowed? Default is \code{FALSE}.
-#' @param empty.ok [\code{logical(1)}]\cr
-#'  Are empty strings allowed? Default is \code{TRUE}.
+#' @inheritParams checkCharacter
 #' @template checker
 #' @family scalars
 #' @export
@@ -17,15 +16,15 @@
 #' @examples
 #' testString("a")
 #' testString(letters)
-checkString = function(x, na.ok = FALSE, empty.ok = TRUE) {
-  .Call(c_check_string, x, na.ok, empty.ok)
+checkString = function(x, na.ok = FALSE, min.chars = NULL, pattern = NULL, fixed = NULL, ignore.case = FALSE) {
+  .Call(c_check_string, x, na.ok, min.chars) %and% checkCharacterProps(x, pattern, fixed, ignore.case)
 }
 
 #' @export
 #' @include makeAssertion.r
 #' @template assert
 #' @rdname checkString
-assertString = makeAssertionFunction(checkString, c.fun = "c_check_string")
+assertString = makeAssertionFunction(checkString)
 
 #' @export
 #' @rdname checkString
@@ -34,7 +33,7 @@ assert_string = assertString
 #' @export
 #' @include makeTest.r
 #' @rdname checkString
-testString = makeTestFunction(checkString, c.fun = "c_check_string")
+testString = makeTestFunction(checkString)
 
 #' @export
 #' @rdname checkString
@@ -44,4 +43,4 @@ test_string = testString
 #' @include makeExpectation.r
 #' @template expect
 #' @rdname checkString
-expect_string = makeExpectationFunction(checkString, c.fun = "c_check_string")
+expect_string = makeExpectationFunction(checkString)