+++ inst/NEWS.md
@@ -1,5 +1,9 @@
 # Version 1.7.2 (2016-xx-xx)
-* New argument `empty.ok` for `*String` family of functions.
+* Added `checkOS()`.
+* Argument `fixed` for `*Character` functions now accepts a string instead of a
+  boolean value and thus can be directly be used for  substring search.
+* New arguments `min.chars`, `pattern`, `fixed` and `ignore.case`  for the
+  `*String` family of functions.
 
 # Version 1.7.1 (2016-02-02)
 * Fixed a segfault while checking an upper bound in qassert/qtest.